GC fastlib_util library and move wildcard_match.h to juniper.
