--- fastlib/src/vespa/fastlib/util/.gitignore
@@ -1,15 +1,0 @@
-*.So
-*.a
-*.ilk
-*.lib
-*.o
-*.obj
-*.pdb
-.cvsignore
-.depend
-.pure
-Debug
-Makefile
-SunWS_cache
-extcase
-extprop
