--- fastlib/src/vespa/fastlib/util/CMakeLists.txt
@@ -1,6 +1,0 @@
-# Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-vespa_add_library(fastlib_util
-    SOURCES
-    INSTALL lib64
-    DEPENDS
-)
