--- fastlib/src/vespa/fastlib/util/tests/.gitignore
@@ -1,42 +1,0 @@
-*.So
-*.a
-*.ilk
-*.o
-*.obj
-*.pdb
-.cvsignore
-.depend
-.pure
-Debug
-Makefile
-SunWS_cache
-arraytest
-bagtest
-base64test
-bitstest
-bloomfilter.*
-bloomfiltertest
-bobhashtest
-crc32test
-datetimetest
-debugtest
-fixedarraytest
-flisttest
-genkeyhashmaptest
-hashmaptest
-hashtabletest
-refcounttest
-resourcestatstest
-ringbuffertest
-scopeguardtest
-singletontest
-stacktest
-stringbuffertest
-stringtest
-stringutilitytest
-trietest
-utiltestsuite
-wildcard_match_test
-fastlib_wildcard_match_test_app
-fastlib_bagtest_app
-fastlib_base64test_app
